recoding all

+++ app/Repositories/AgeRepository.php
@@ -4,6 +4,7 @@ namespace App\Repositories;
 
 use App\Age;
 use Validator;
+use Exception;
 
 class AgeRepository implements RepositoryInterface
 {
@@ -16,7 +17,11 @@ class AgeRepository implements RepositoryInterface
 
     public function getById($id)
     {
-        return $this->modelClass::find($id);
+        if ($this->verifyIfExists($id)) {
+            return $this->modelClass::find($id);
+        } else {
+            throw new Exception("Record not found");
+        }
     }
 
     public function create(array $data)
@@ -27,7 +32,7 @@ class AgeRepository implements RepositoryInterface
         ];
         $validator = Validator::make($data, $rules);
         if ($validator->fails()) {
-            return response()->json($validator->errors(), 400);
+            throw new Exception($validator->errors());
         }
         return $this->modelClass::create($data);
     }
@@ -40,15 +45,31 @@ class AgeRepository implements RepositoryInterface
         ];
         $validator = Validator::make($data, $rules);
         if ($validator->fails()) {
-            return response()->json($validator->errors(), 400);
+            throw new Exception($validator->errors());
         }
-        $this->modelClass::where('id', $id)->update($data);
-        return $this->modelClass::find($id);
+        if ($this->verifyIfExists($id)) {
+            $this->modelClass::where('id', $id)->update($data);
+            return $this->modelClass::find($id);
+        } else {
+            throw new Exception("Record not found");
+        }
     }
 
     public function delete($id)
     {
-        $data = $this->modelClass::find($id);
-        return $data->delete($id);
+        if ($this->verifyIfExists($id)) {
+            $data = $this->modelClass::find($id);
+            return $data->delete($id);
+        } else {
+            throw new Exception("Record not found");
+        }
+    }
+
+    private function verifyIfExists($id)
+    {
+        if (is_null($this->modelClass::find($id))) {
+            return false;
+        }
+        return true;
     }
 }
